+++ elements/src/components.d.ts
@@ -254,7 +254,8 @@ export namespace Components {
     interface PosSubjects {
     }
     /**
-     * A tool to manage attachments of a thing.
+     * A tool to manage attachments of a thing!
+     * @since 0.35.0
      */
     interface PosToolAttachments {
     }
@@ -1088,7 +1089,8 @@ declare global {
         new (): HTMLPosSubjectsElement;
     };
     /**
-     * A tool to manage attachments of a thing.
+     * A tool to manage attachments of a thing!
+     * @since 0.35.0
      */
     interface HTMLPosToolAttachmentsElement extends Components.PosToolAttachments, HTMLStencilElement {
     }
@@ -1576,7 +1578,8 @@ declare namespace LocalJSX {
         "onPod-os:resource"?: (event: PosSubjectsCustomEvent<any>) => void;
     }
     /**
-     * A tool to manage attachments of a thing.
+     * A tool to manage attachments of a thing!
+     * @since 0.35.0
      */
     interface PosToolAttachments {
     }
@@ -1759,7 +1762,8 @@ declare module "@stencil/core" {
             "pos-setting-offline-cache": LocalJSX.PosSettingOfflineCache & JSXBase.HTMLAttributes<HTMLPosSettingOfflineCacheElement>;
             "pos-subjects": LocalJSX.PosSubjects & JSXBase.HTMLAttributes<HTMLPosSubjectsElement>;
             /**
-             * A tool to manage attachments of a thing.
+             * A tool to manage attachments of a thing!
+             * @since 0.35.0
              */
             "pos-tool-attachments": LocalJSX.PosToolAttachments & JSXBase.HTMLAttributes<HTMLPosToolAttachmentsElement>;
             /**